requirements.txt
certifi==2021.10.8
pulsar-client==2.9.1
six==1.16.0
